Create format target
With the addition of Prettier in [1] there is now the possibility of
automatically formatting source code using a tool. Following this, the
`make format` target has been added to run all available formatters.
Like lint targets, there will be a separate target for every type of
formatter (so far only JavaScript).

This new make target has also been documented in the Contributing
Guidelines to make it easier for contributors to discover this command.

--
1. 275dd39209ab3c53d6d67e880b71c10c73458157