Support building with Podman
Update the Makefile to support building the scanner image with Podman[1]
instead of Docker[2]. The aim being to support both engines so users
will be able to choose which one they want to use. To that end, the
continuous integration job that builds the image has been updated to
build with both Docker and Podman (the egress policy is set to audit to
find out what endpoints the `podman` variant needs).

The Contributing Guidelines have been updated accordingly as well.

More "support" for Podman, such as running the tests with Podman, may be
added later. It may also make sense to change some of the terminology
away from "docker" (e.g. the `audit-docker` target), again this is
considered a task for later.

--
1. https://podman.io/
2. https://www.docker.com/