Document reason each allowed license is allowed
Include reasoning for allowing every allowlisted license in the project
to:
1. make intent clearer, and
2. allow for more constructive public review.

This follows, in particular, recent concern about alowing `GPL-2.0-only`
and `GPL-2.0-or-later` in this project by the commit author. The reason
given here differs from the original reason in [1] and is based on the
reasoning given in [2].

For now the reasoning is included in the licensee configuration as it's
the single source of truth for OSS license compliance in this project.
licensee will happily ignore the additional field. Nevertheless, the
decision it put the reasons here may well be reconsidered in the future.

--
1. d3041af6ab30ebcac6a49d5a7f9895defb3c9730
2. https://opensource.com/article/18/1/containers-gpl-and-copyleft