Run tests using Podman
Add support for running tests with Podman [1] as the container engine
instead of Docker [2]. Similar to building, you can run:

    make test                # test with Docker
    make test ENGINE=podman  # test with Podman
    make test ENGINE=docker  # also test with Docker

The primary goal being to support running tests with both engines in CI.
But, conveniently, it also allows for testing with either engine or
testing if Docker isn't available on a contributor's system.

The Contributing Guidelines have been updated to include instructions
for how to run the tests using Podman, as well as how to update the
snapshots using Podman.

--
1. https://podman.io/
2. https://www.docker.com/