Separate Docker and Podman image reference files
Update the temporary image reference file logic (used to determine if
the image needs to be rebuild) to create separate files for Docker and
Podman images so as to avoid (not) rebuilding when different container
engines are interleaved.

Also, rename dirty variables to be container engine-agnostic.

Note: the `ENGINE?=` definition must appear before, in this case,
`IMAGES_DIR` as otherwise it's undefined and an empty string will be
used instead, invalidating the engine-based reference file separation.