Use old target for auditing the `v0` ref
Following [1] the Make target for auditing the container image is now
different between the `main` branch ref and the `v0` ref. As such, the
nightly auditing job fails because the executed target doesn't exist at
the `v0` ref. To solve this, the command to audit the container image is
made conditional on the ref.

--
1. 91526b743defce89e2283ca1f04fa8e69a6ec40b