Add unnecessary license correction check
Implement a check in the `check-licenses.js` script to look for license
corrections that are not necessary. I.e., the SBOM contains the same
license information as the correction.

To avoid keeping around unnecessary corrections the script has been
reworked to exit with a non-zero exit code if this is the case, thus
blocking continuous integration and forcing a resolution.