Base corrections on id rather than name
Update the `check-licenses.js` script to base corrections on the id of
a dependency instead of the name. The id (hopefully) is unique as opposed
to the name. Besides different dependencies from different ecosystems
possibly having the same name, in this case "busybox" is found twice
(once by `binary-cataloger` and once by `apkdb-cataloger`) and
is thus present in the SBOM twice. Moreover, one of these finds the
license for "busybox" while the other does not, thus leading to a
"unnecessary correction(s) detected" warning.