Add and use `.nvmrc`
Create a file called `.nvmrc` for configuring the Node.js version used
for development. This file originally is a configuration file for a
tool called nvm [1], but has grown and is more widely adopted. Including
by the `actions/setup-node` Action [2].

Using this file aids local development (as it makes it easier to use the
expected Node.js version if the contributor has relevant tooling) and
slightly simplifies updating the Node.js version for development as
fewer files have to be changed.

--
1. https://github.com/nvm-sh/nvm
2. https://github.com/actions/setup-node