Use fixed Node.js when auditing the v0 ref
Since [1] the project uses the `.nvmrc` file to record the Node.js
version used for development as well as in Continuous Integration jobs.
This file is not  yet present on at the `v0` ref, which is audited
nightly, and thus may fail. This aims to prevent that failure by
changing the CI job behaviour for the `v0` ref only.

This change should be reverted/removed when the `.nvmrc` file lands in
the `v0` ref.

--
1. dc620b6f17171813d6c0ff43325375df71fb0a7b