Rescope corrections to `(name, foundBy)`
Update how corrections are applied from using the `id` to using the
`name`+`foundBy` values combined. Unfortunately, the `id` seems to vary
in some cases (observed: different host operating system) so isn't
suitable for corrections. Using the `foundBy` value (in conjunction with
the `name`) addresses the concern which originally [1] invoked the
switch to `id`-based corrections. Namely, if an artifact is `foundBy`
multiple catalogers, the correction only applies to a specific
cateloger's findings.

--
1. e0f17f6deec8958b17247d5a7291643d4e9c58d7