Update whole word matching logic in image licenses check script
Refactor the script for checking image licenses with a new whole word
match logic. The old logic was fine, but unfortunately relied on string
interpolation into a Regular Expression which is not ideal - it's a
potential attack vector. Granted, there is no real attack vector here
[^1], but I'm changing it primarily to avoid anyone looking at this code
copying the vulnerable implementation - and partly to prove I can :)

--
^1: The first and last parts are declared just prior to the RegExp
    instantiation, the middle part always comes from `allowedLicenses`
    which is recorded elsewhere in the codebase (in `.licensee.json`
    to be specific).