Use SPDX schema SBOM for image auditing
Update the Syft configuration to generate both a Syft schema SBOM and
SPDX schema SBOM (more on why both below) and switch Grype from using
the former to using the latter. I believe this will make things more
stable and avoid auditing problems when Syft's SBOM format has changed
but Grype hasn't been updated yet.

The Syft schema SBOM is kept, for now, because the `check-licenses.js`
script is coded for that schema and doesn't work with the SPDX schema.
The script may be updated to work with the SPDX schema instead in the
future, but this is considered out of scope for this commit.