Remove `.dockerignore`
Replace the use of `.dockerignore` by explicitly copying files into the
image. This change is to be more container engine agnostic and is
feasible for this project as there's only three files that need to be
copied into the container.

As far as I can tell Docker does not support an alternative name for the
ignore file. Podman does have a `--ignorefile` option, but when used
with Docker it complains "unknown flag: --ignorefile".