Add basic documentation tests
Implement a second test suite that is aimed at testing scripts in the
documentation. It iterates over the files in the repository looking for
MarkDown files. For every MarkDown file it will try to run each script
(i.e. a snippet with the language set to "shell"), leveraging snapshot
testing to ensure the result doesn't change (since a negative result may
be expected).

The system features a lightweight directive system to ignore files or
individual scripts. This is currently leveraged to ignore scripts that
1) mutate the state of the repository (in README.md), 2) will lead to
infinite recursion (in CONTRIBUTING.md), or 3) is otherwise dangerous or
uninteresting (in RELEASE.md).