Migrate to ESLint flat config
This updates the configuration for ESLint used by this plugin from the
legacy format of ESLint v8 and earlier to the new "flat config" [1]. The
migration is based on [2] as well as prior experience migrating other
projects. The migration itself is mostly straightforward now that it has
been implemented. However, there are some additional changes, in
particular to the `Containerfile`.

The `Containerfile` changes account for general new behavior of ESLint
v9. In particular, some CLI options have changed (only renames for the
purposes of this project), the way file (extensions) are specified has
changed (moved from the CLI to the config file), and the target dir as
well as ignore scope has changed. The latter resulted in the addition
of the `WORKDIR /project` directive just before the `ENTRYPOINT` to
make ESLint think the current working directory is `/project` s.t. it
will treat it as the directory to scan. Correspondingly, the CLI is now
invoked with absolute paths and the final ESLint argument is just `.`
(i.e. current dir).

--
1. https://eslint.org/blog/2022/08/new-config-system-part-1/
2. https://eslint.org/docs/latest/extend/plugin-migration-flat-config