Use Ubuntu 24.04 for jobs with Podman
This is a followup to 4f88d3cc676de0072c3b3d66a9c0d7d8db005f35 to
upgrade the remaining CI jobs (those involving Podman) to Ubuntu 24.04.
These jobs encountered problems when upgrading initially, which were
put aside to be able to upgrade most jobs asap with this as a followup.