Replace gitleaks by trufflehog
Replace the use of gitleaks for secret scanning by trufflehog (in
addition to GitHub). The motivation for this switch is that gitleaks is
poorly maintained and keeps using deprecated APIs. Unfortunately
trufflehog uses a Docker image making the action technically unpinnable,
but I think this trade-off is worth it here.