Use `depreman` to manage deprecations
Depreman is a tool to manage deprecations. I'm adding it here to get
insight and control over deprecated dependencies of the project - and
also to dogfeed it on my own projects. It is currently used to ignore
several deprecation warnings transitively introduced through AVA that
have an open issue tracking the warning.

This also upgrades the dependency `licensee` to the latest version in
order to resolve deprecated transitive dependencies.

The license for `depreman` is explicitly allowlist in the `licensee`
config because it is distributed under the AGPL license. This is fine
for this devDependency, but would not be okay for "runtime" deps, hence
the exception.