Switch to the Node.js test runner
Replace AVA in favor of the Node.js test runner to reduce the number of
distinct dependencies^1 as well as try out the Node.js test runner's
snapshot testing support. The switch is relatively easy, the majority of
the changes are due to the addition of a top-level `test()` used for the
grouping with the Node.js test runner (which, unlike AVA, doesn't group
by test file) and differences in the snapshot format.

--
^1: as well as cut down on deprecations in the dependency tree.