Check container reproducibility with diffoci
Create a Makefile target to test the reproducibility of the project
container using diffoci[1]. This tool provides a way to test for
reproducibility of OCI containers by virtue of being able to display
the diff between two versions of a container. It also allows for
omitting some uninteresting diffs such as timestamps (with the
`--semantic` flag).

This is an early prototype as the setup currently introduces the Go
programming as a dependency and has non-optimal or missing version
pinning.

--
1. https://github.com/reproducible-containers/diffoci