Use `--omit dev` to audit deprecations on releases
Update the `audit-deprecations-npm` command to accept an ARGS argument,
just like the `audit-vulnerabilities-npm` command. Use this to append
the `--omit dev` option when auditing the latest release for deprecation
warnings in CI.

This CI change should not break the nightly check for deprecation
warnings on v0 because prior to this commit the make target would just
ignore the `ARGS`. Only when this change lands in a release, the latest
release will be audited for deprecation warnings excluding development
dependencies.