Ignore container vulns
Ignore three vulnerabilities detected by Grype that don't affect this
project.

`CVE-2024-13176` is related to cryptography, which is not used at
runtime. `CVE-2025-23083` and `CVE-2025-23090` are related to the Node
permission system, which we don't use at all.

Also snuk in two corrections for auditing deprecation warnings. 1 is to
hide the audit command in make and 1 is to correctly omit warnings from
development dependencies (`depreman` does not support the previous
syntax).