Refactor CI auditing
This changes how we audit (for vulns and deprecations) in CI from a
reusable workflow to two self-contained workflows. I believe this makes
the setup a bit more explicit (when each happens) at the cost of some
duplication (which is counteracted by the removal of conditional logic)
and potential audit failures if there are build failures (because there
is no longer a `needs:` relation between the build job and audit jobs,
tho at the same time only one of the audit jobs should actually be
blocked by the build job).