Patch publish workflow
Correct the publish workflow jobs for pushing to both Docker Hub and
GHCR. Both failed due to an incorrect full name for the image. For
Docker Hub this is because the `matrix.registry` value is an empty
string and for GHCR this is because the name must be
`ericcornelissen/js-re-scan` (with two 'c's instead of one).