Update the publishing workflow
Change the publish workflow to fix some semantic issues as well as
publishing to GHCR. The former are just some naming corrections. The
latter is achieved by 1) not trying to log in to GHCR (because the
runner's `GITHUB_TOKEN` _should_ be used), and 2) adding the
`packages: write` (to allow the runner's `GITHUB_TOKEN` to push the
image).