Trigger dev auditing based on check workflow
Change the triggers of `audit-dev.yml` from pushes and pull requests to
instead depend on the "Check" workflow (`check.yml`). This is because
the auditing can only succeed if the project can be build, so this
dependency ensures these jobs are only run if the build has been
checked. This comes at the cost of the path filters, meaning the
auditing may now occur when no relevant files have changed.