Add "Trophies"
Add a list of bugs and advisories found using this scanner as a list of
"Trophies" (a practice I first found out about from `dvyukov/go-fuzz`).
This helps show the usefulness of the project as it indicates it can
find problems in real-world projects.

Also include all trophies in the tests as individual cases.