Set up automated GitHub Release creation with the `gh` CLI
Set up GitHub Release creation leveraging the by-default installed `gh`
CLI in GitHub Actions. Creating a release on GitHub enables the use of
immutable release which helps as a defense in depth strategy against
supply chain attacks (even if users are unlikely to use the git tags for
this project).